feat: works

+++ internal/entity/account.go
@@ -30,7 +30,7 @@ func (a *AccountID) Scan(src any) error {
 		return err
 	}
 
-	value, ok := iv.(int)
+	value, ok := iv.(int64)
 	if !ok {
 		return errors.New("failed to scan account id")
 	}
